flask/requirements.txt
flask
uwsgi
python-dotenv